Create common core for data and knowledge service interfaces, add sorter for repeatable abilities
